Replaces balance context with subgraph context Query builder is added
