+++ src/context/index.tsx
@@ -1,5 +1,5 @@
-import { useOnboard } from './onboard.context';
-import { useBalances } from './balance.context';
-import { useLiquidityPositions } from './positions.context';
+import { useOnboard } from '@/context/onboard.context';
+import { useSubgraph } from '@/context/subgraph.context';
+import { useLiquidityPositions } from '@/context/positions.context';
 
-export { useOnboard, useBalances, useLiquidityPositions };
+export { useOnboard, useSubgraph, useLiquidityPositions };
