+++ src/lib/claim.ts
@@ -5,7 +5,9 @@ import config from '../configuration';
 import { abi as MERKLE_ABI } from '../artifacts/MerkleDrop.json';
 import { abi as TOKEN_DISTRO_ABI } from '../artifacts/TokenDistro.json';
 import { TransactionResponse, Web3Provider } from '@ethersproject/providers';
-import { fetchBalances } from '@/services/subgraph';
+import { fetchSubgraph } from '@/services/subgraph.service';
+import { SubgraphQueryBuilder } from '@/lib/subgraph/subgraphQueryBuilder';
+import { transformSubgraphData } from '@/lib/subgraph/subgraphDataTransform';
 
 export const fetchAirDropClaimData = async (
 	address: string,
@@ -35,11 +37,18 @@ export const fetchAirDropClaimData = async (
 };
 
 export const hasClaimedAirDrop = async (address: string): Promise<boolean> => {
-	const { givDropClaimed } = await fetchBalances(
-		config.XDAI_NETWORK_NUMBER,
-		address,
-	);
-	return givDropClaimed;
+	try {
+		const response = await fetchSubgraph(
+			SubgraphQueryBuilder.getXDaiQuery(address),
+			config.XDAI_NETWORK_NUMBER,
+		);
+		const { balances } = await transformSubgraphData(response);
+		return balances.givDropClaimed;
+	} catch (e) {
+		console.error('Error on fetching subgraph');
+		// It's better to continue givDrop if subgraph is unreachable
+		return false;
+	}
 };
 
 export const claimAirDrop = async (
